2.4.Arrow Functions{Arrow Functions Everywhere! }
We can finally remove this silly var = self thing. And instead, below, use this. But because we're inside of another anonymous function, replace it with the new arrow syntax to get things work

Arrow Functions Everywhere!

Let's use the arrow syntax everywhere! Below, when we use SweetAlert, remove the self variable and - in preConfirm - use () for empty arguments, then add the arrow. Inside, we can use this! Use the arrow function again below.
Here, we're not using this, but I like to stay consistent and use the arrow function everywhere.
Keep going! Inside the next method, remove self, and add our arrow function
Do the same for fadeOut(). But here, we were using this, which previously pointed to the DOM Element object that was fading out. We can't use this anymore, but that's fine! Replace it with .remove() and then this.updateTotalWeight()
Double-check that things work. Refresh! Delete one of the items and... perfect!
Since we're going to use arrow functions for all anonymous functions, search for function(). Yep, we're going to replace everything, except for the methods in our objects. Remove function(), and add the arrow
Repeat it again, and remove another self variable: just use this
I'll fast-forward through the rest of the changes.