4.1.var Versus let: Scope!
When you look at ES2015 code, one thing tends to jump out immediately: suddenly instead of seeing var everywhere... you see something called let! In our JS file, if you scroll down to the bottom, PhpStorm has highlighted my var with a warning:

  var used instead of let or const

What's going on?
To find out, change var to let and then console.log(totalWeight)
Now, go refresh. This function is called a bunch of times as the table is loading. And... it looks like everything works fine. It looks like var and let are equivalent?

In most cases, that's true! let is a new way to initialize a variable that's almost the same as var. 99% of the time, you can use either one, and it won't make a difference.