4.2.var Versus let: Scope! {Understanding the Scope of var}
Understanding the Scope of var

So... what about that last 1%? Well, it has to do with variable scope. To understand, go back to play.js. At the top, add something silly: if (true) then aGreatNumber = 42, which of course, is a great number.
When we run it, it re-assigns the variable to 42. No surprises.