4.3.var Versus let: Scope! {Understanding the Scope of var}
Understanding the Scope of var

But what if we added var aGreatNumber = 42?. I shouldn't need to say var again: the variable has already been initialized. But, will this give us an error? Or change anything?

Let's find out! No! We still see 42. When we use the second var, it re-declares a new variable called aGreatNumber. But that doesn't make any real difference: down below, it prints the new variable's value: 42.