4.4.var Versus let: Scope! {Understanding the Scope of var}
Understanding the Scope of var

But now, wrap this same code in a self-executing function. Use the new arrow syntax to be trendy, then execute it immediately
Will this change anything? Try it!
Woh! It prints as 10! Why!
Remember, the scope of a variable created with var is whatever function it is inside of. Let's follow the code. First, we create the variable and set it to 10. Then we create a new variable set to 42. But since this is inside of a function, its scope is only this function. In other words, inside of the self-executing block, aGreatNumber is 42. But outside, the original variable still exists, and it's still set to 10. Since we're printing it from outside the function, we see 10.
Okay okay, I know, this can be confusing. And most of the time... this subtle scope stuff doesn't make any difference. But, this is exactly where var and let different.