4.5.var Versus let: Scope! {The "Block" Scope of let}
The "Block" Scope of let

Let me show you. Remove the self-executing function and change each var to let. If let and var behaved exactly the same, we would expect this - just like before - to print 42. Try it.

But no! It prints 10! And this is the difference between var and let. With var - just like with any variable in PHP - a variable's scope is the function it's inside of, plus any embedded functions. But let is different: it's said to be "block-scoped". That means that anytime you have a new open curly brace ({) - like an if statement or for loop - you've entered a new scope for let. In this case, let is equal to 42, only inside of the if statement. Outside, it's a completely different variable, which is set to 10.