5.4.var Versus let: Hoisting! {Hello Mr Variable Hoisting}
Hello Mr Variable Hoisting

The reason for this is something called variable hoisting, a term you'll see a lot around JavaScript... I think mostly because it has a cool name. It's actually not that important, but I want to tell you a little bit about it so you don't have to worry about it ever again.

In PHP, we never need to initialize a variable with a special keyword. We don't say var $aGreatNumber = 10, we just say $aGreatNumber = 10 and we're good to go. But in many other languages, including JavaScript, you must initialize a variable first with a keyword.

When you use var to initialize a variable, when JavaScript executes, it basically finds all of your var variables, goes to the top of that variable's scope - usually the top of whatever function it's inside of, but in this case, it's the top of the file - and effectively does this: var aGreatNumber. That initializes the variable, but doesn't set it to any value. This is called variable hoisting: and it's the reason that we get undefined instead of an error when we try to use a variable that's declared with var... before it's declared.