5.5.var Versus let: Hoisting! {Hello Mr Variable Hoisting}
But when we change this to let, we already saw that this does throw a ReferenceError. And that's kinda great! I mean, isn't that what we would expect to happen when we reference a variable that hasn't been created yet!

So with var, variables are hoisted to the top. But with let, that doesn't happen, and that's kinda cool. Well, technically, let also does variable hoisting, but thanks to something called the "temporal dead zone" - also an awesome name - let acts normal: as if its variables were not hoisted.