6.2.const Versus let{You can Change a const?}
But if we comment-out the line where we change the variable, it works just like we expect. As far as scope goes, const and let work the same. So really, const and let are identical... except that you can't modify a const variable.

You can Change a const?

Well actually, that's not completely accurate. Create another const variable called aGreatObject. Inside it, set withGreatKeys to true. Below, change that object: aGreatObject.withGreatKeys = false. Log that object at the bottom
Will this work?

Try it! It does work! The withGreatKeys property did change! Here's the truth: when you use const, it's not that the value of that variable can't change. The object can change. Instead, using const means that you cannot reassign the aGreatObject variable to something else in memory. It must be assigned only once, to this object. But after that, the object is free to change.