6.3.const Versus let{var, let and const}
var, let and const

Phew! Okay, in ES2015, var, let, and const do have some differences. But a lot of the time, you will see people use one or the other purely based on personal preference, or whatever is trendy that week. The differences are minor.
In our case, because you know I love to write hipster code, let's change each let to const. Start at the top: const $link makes sense. We don't need to reassign that. The same is true for deleteUrl, $row, $form, and formData
Sure, we modify formData, but we know that's ok with const!
Keep going: $form, fieldName, $wrapper, $error... and eventually we get to the last one: totalWeight. But this variable can't be set to a const: we set it to 0, but then reassign it in each loop. This is a perfect case for let
Let's take our app for a spin! Refresh! And try deleting something. Woohoo! Yep, you can pretty much choose between var, let and cat, I mean, const.