7.2.Object Literals & Optional Args{Object Keys... without the Key}
Well, in ES2015, if your key and your value are the same, you can just leave off the key. Yep, this means the same thing as before. So if you suddenly see an associative array or object where one of its keys is missing... well, it is the key... and the value.