7.3.Object Literals & Optional Args{Short Method Syntax}
Short Method Syntax

Next, you can do something similar with methods inside an object. The loadRepLogs() method is just a loadRepLogs key assigned to a function. Simple, but too much work, maybe? In ES2015, we can shorten this to loadRepLogs(). So much cooler! Let's change it everywhere! Search for the word function, because almost everything is about to change. Ultimately, the only function keywords that will be left are for the self-executing function - which could be an arrow function - and the two constructors.
Nice