7.4.Object Literals & Optional Args{Optional Args}
Optional Args

Ready for one more cool thing? This one is easy. Suppose we have a new method, not calculateTotalWeight(), but getTotalWeightString(). Use the new shorthand syntax and return this.calculateTotalWeight() and append "pounds" to it
Perfect! Then above, in updateTotalWeightLifted(), instead of calling calculateTotalWeight() and passing that to .html(), pass getTotalWeightString()
Ok, nothing too crazy so far: when we refresh, at the bottom, yep, "pounds".