7.5.Object Literals & Optional Args{Optional Args}
But now suppose that we want to set a max weight on that. What I mean is, if we are over a certain weight - maybe 500 - instead of printing the actual total, we want to print "500+"
Start by adding a new argument called maxWeight. Then say let weight = this.calculateTotalWeight(). And if weight > maxWeight, add weight = maxWeight + '+'. At the bottom, return weight and "pounds". Head up top to try this: when we call getTotalWeightString(), pass 500
Refresh! It works! We see the 500+ at the bottom.