7.6.Object Literals & Optional Args{Optional Args}
But what if I wanted to make this argument optional with a default value of 500? You could do this before in JavaScript, but it was ugly. Now, thanks to our new best friend ES2015, we can say maxWeight = 500 - the same way we do in PHP
Now, we can remove the argument and everything is still happy!
So, yay! Finally, JavaScript has optional function arguments! And a second yay, because we are ready to learn perhaps the biggest change in ES2015: JavaScript classes.