8.3.Legit JavaScript Classes{Classes do not have Properties}
Properties are not supported inside classes: only methods are allowed. That may seem weird - but it'll be more clear why in a minute. For now, change this to be a method: _getSelectors(). Add a return statement, and everything is happy. Well, everything except for the couple of places where we reference the _selectors property. Yea, this._selectors, that's not going to work
But don't fix it! Let's come back in a minute.