8.5.Legit JavaScript Classes{Magic get Methods}
Magic get Methods

Time to go back and fix this _getSelectors() problem. The easiest thing would be to update this._selectors to this._getSelectors(). But, there's a cooler way.
Rename the method back to _selectors(), and then add a "get space" in front of it
Woh! Instantly, PhpStorm is happy: this is a valid syntax. And when you search for _selectors, PhpStorm is happy about those calls too!
This is the new "get" syntax: a special new feature from ES2015 that allows you to define a method that should be called whenever someone tries to access a property, like _selectors. There's of course also a "set" version of this, which would be called when someone tries to set the _selectors property.
So even though classes don't technically support properties, you can effectively create properties by using these get and set methods.
Oh, and btw, just to be clear: even though you can't define a property on a class, you can still set whatever properties you want on the object, after it's instantiated
That hasn't changed.
Ok team! Try out our app! Refresh! It works! Wait, no, an error! Blast! It says:

  RepLogApp is not defined

And the error is from our template: app/Resources/views/lift/index.html.twig
Ah, this code is fine: the problem is that the RepLogApp class only lives within this self executing function:

 206 lines  web/assets/js/RepLogApp.js
'use strict';
(function(window, $, Routing, swal) {
    class RepLogApp {
    }
})(window, jQuery, Routing, swal);
It's the same problem we had in the first episode with scope.