10.1.Class Inheritance and super Calls
In PHP, a class can extend another class. So, can we do that in JavaScript? Totally! And once again, you're going to love it.
To try this out, let's go back to the play.js file. Create a new class: AGreatClass, because, it's going to be a great class. Give it a constructor with a greatNumber arg and set that on a property.
Below, add one new method: returnGreatThings, which it will, because it's going to return our greatNumber!
Finally, create a new constant, aGreatObject, set to new AGreatClass(42). Let's console.log(aGreatObject.returnGreatThings())
There's no mystery about what this is going to return. Run the file! Yep, 42!