10.2.Class Inheritance and super Calls{Extending a Class}
Extending a Class

Now, let's create another class, class AnotherGreatClass, because we're on a roll. But, I want this to be a sub class of AGreatClass. How? The same way we do it in PHP: extends AGreatClass. That is it. We're not overriding anything yet, but this should work. Change the variable to be a new AnotherGreatClass, and then run the file!
Success