10.3.Class Inheritance and super Calls{Overriding a Method and Calling super}
Overriding a Method and Calling super

From here, pretty much all the rules of extending things in PHP work in JavaScript, with just a little bit of a different syntax. For example, we can override returnGreatThings(). Let's return something else that's great: adventure!
This completely overrides the parent class method.