10.5.Class Inheritance and super Calls{Overriding the constructor}
Overriding the constructor

Can we override the constructor in the same way? Because really, I should be able to pass the greatNumber and the great thing, adventure, into this class.

Override the constructor and give it just one argument: greatWord. Then, set this.greatWord = greatWord

Before we try to print that below, what do you think will happen when we run this? We're setting the greatWord property... but we're not calling the parent constructor. Actually, try to run it!

Ah, we get an error in the constructor!

  ReferenceError: this is not defined

Interesting! Inside construct(), there is no this variable? And PhpStorm is trying to tell us why:

  Missed superclass's constructor invocation