11.4.Destructuring{Destructuring & Default Values}
If you think this might be possible, you can give that variable a default, like whatever
Now, if the key doesn't exist, it'll get set to whatever.

So, this is destructuring. It may or may not be useful to you, but you will see it! Don't let it surprise you!