12.2.The... Spread Operator
But of course, there are a lot of yummy things in the world, and since I'm from the US, my yummyThings should probably have a cheeseburger

What will happen now? The array has 4 things, but we only have 3 arguments? Let's find out! Run the script!

It's the same result! Thanks to the spread operator, pizza is passed as the first argument, then gelato, sushi and finally, cheeseburger is passed as the 4th argument. Since the function doesn't have a 4th argument, it's just ignored!