12.5.The... Spread Operator{Spread Operator for Creating a new Array}
So how could we make copyOfGreatThings a true copy? The spread operator has an answer: re-assign copyOfGreatThings to [...greatThings]
And that is it! This will create a new array, and then put each item from greatThings into it, one-by-one.
Try it! Yes! We can see summer in the copy, but we did not modify the original array.
Ok, let's move on something that's instantly useful: template strings!