13.1.Template Strings
You know what's always kind of a bummer... in any language? Creating strings that have variables inside. I know, that doesn't sound like that big of a deal, but it kind of is... especially in JavaScript!

For example, let's say that our favorite food is, of course, gelato.
And below that, we want to create a string that talks about this: The year is, end quote, plus, and then new Date().getFullYear(), another plus, open quote, space, my favorite food is, another quote, one more plus, and favoriteFood.
Perfectly straightforward and perfectly ugly! At the bottom, log that. This is the way we used to do things in JavaScript.