13.2.Template Strings{ES2015: Nicer Strings})(window, jQuery, Routing, swal);
ES2015: Nicer Strings!

Thanks to ES2015, we now have something called Template Strings. And it's awesome. Instead of quotes, use a tick. And as soon as you do that, you're allowed to put variables inside of your string! Remove this single quote plus garbage. Replace it with ${, the variable name, then }. Then, remove all the ugliness!
Usually, you'll use this to print a variable, but you can use any expression you want: like we just did with the date.