13.3.Template Strings{Template Strings and Line Breaks}
Template Strings and Line Breaks

There's something else that makes strings especially annoying in JavaScript: line breaks. In the first episode, we created a client side template and used Underscore.js to render it. We decided to put the template itself into Twig app/Resources/views/lift/index.html.twig. Then, in _addRow(), we found that script element by its id and fetched the string
So, why did we put the the template in Twig? Well, it's not a bad option, but mostly, we did it because putting that template into JavaScript was, basically, impossible.
Why? Try it! Copy the template, find the bottom of RepLogApp.js, create a new const rowTemplate and paste the string inside single quotes
Oh wow, PHP storm is SO angry with me. Furious! And that's because, unlike PHP, you are not allowed to have line breaks inside traditional strings. And that makes putting a template inside of JavaScript basically impossible.