13.5.Template Strings{Using Template Strings as a Template Engine}
Using Template Strings as a Template Engine

But let's go a step further! Now that we have this nice ability to embed variables into our strings, could we create our own templating engine, instead of using the one from Underscore? I mean, there's nothing wrong with Underscore, but we can accomplish this with pure template strings.
First, update the template with ${repLog.totalWeightLifted}. In a moment, instead of making the individual variables available, we'll just set one variable: the repLog object.
Of course, PhpStorm is angry because there is no repLog variable... but we can fix that!
Next, make the same change very carefully to the other variables: replacing the kind of ugly syntax with the template string syntax. And yea, be more careful than I am: I keep adding extra characters!
At this point, this is a valid template string... with one screaming problem: there is no repLog variable! If we refresh now, we of course see:

repLog is not defined