13.6.Template Strings{Turning our Template into a Function}
Turning our Template into a Function

Here's the goal: inside of _addRow(), we need to somehow use that template string, and pass it the repLog variable we have here. But... how can we make this variable available... way down here?
Hey! Here's an idea: let's wrap this in a function! Instead of rowTemplate simply being a string, set it to an arrow function with a repLog argument
And suddenly, the template string will have access to a repLog variable.
Back in _addRow(), remove all this stuff and very simply say html = rowTemplate(repLog)
And that is it! Try that out! Refresh! The fact that it even loaded proves it works: all of these rows are built from that template.