15.2.Map and WeakMap{The Map Object}
In ES2015, we now have a new tool: instead of creating a simple object, we can create a new Map object. The syntax is slightly different: instead of foods.italian = 'gelato', use foods.set('italian', 'gelato').
Repeat this for the other two keys. And at the bottom, fetch the value with foods.get('italian').
Simple and beautiful! And it works exactly like before!