15.3.Map and WeakMap{The Map Object}
Great! So... we have a new Map object... and it's a different way to create an associative array. But why would we use it? Because it comes with some nice helper methods! For example, we can say foods.has('french')

And that returns false. Bummer for us.
It wasn't too difficult to check if a key existed before, but this feels clean.