15.6.Map and WeakMap{Introducing WeakMap... a worse Map?}
Introducing WeakMap... a worse Map?

ES2015 also gives us a very similar new object: WeakMap
And this is where things get a little nuts. Why do we have a Map and a WeakMap?
Let's find out! First try to run our code with WeakMap.
Woh, it explodes!

  Invalid value used as week map key

Map and WeakMap are basically the same... except WeakMap has an extra requirement: its keys must be objects. So yes, for now, it seems like WeakMap is just a worse version of Map.