16.3.Private Variables & WeakMap{Multiple Instances with Map})(window, jQuery, Routing, swal);
Multiple Instances with Map!

Ok, so why not use our cool new Map object to store a collection of Helper objects? let HelperInstances = new Map(). In the constructor(), set the new object into that map: HelperInstances.set()... and for the key - this may look a little weird - use this.
In other words, we key this HelperInstance to ourselves, our instance. That means that later, to use it, say HelperInstances.get(this).getTotalWeightString().
This is awesome! Helper is still private, but now each RepLogApp instance will have its own instance of Helper in the Map.

Just to prove this is not breaking everything, refresh! Woohoo