16.4.Private Variables & WeakMap{Playing with Garbage Collection}
Playing with Garbage Collection

Time for an experiment! Go all the way to the bottom of the file. Create a new RepLogApp object... and just pass in the body tag. Copy this and repeat it three other times
Notice that these are not being used: I'm not setting them to a variable. In other words, they are created, and then they're gone: no longer referenced by anything. Below that - and this won't make sense yet, call setTimeout(), pass it an arrow function, and inside, console.log(HelperInstances). Set that to run five seconds after we load the page.
Ok, refresh! And then wait a few seconds... we should see the Map printed with five Helper objects inside. Yep, we do! One Helper for each RepLogApp we created.