17.4.Array, Set and ES2016{Calculating the Total Weight}
Calculating the Total Weight

But... why are we keeping track of the repLogs? Because now, we can more easily calculate the total weight. Before, we passed the Helper object the  element so that it could find all the tr elements and read the weight from them. We can simplify this! Instead, pass it our Array: this.repLogs
At the bottom of this file, change the constructor() for Helper to have a repLogs argument. Set that on a repLogs property.
Below in calculateTotalWeight(), instead of using the  to find all the tr elements, just pass this.repLogs to the static function. Inside of that, update the argument to repLogs.
Previously, _calculateWeights() would loop over the  and read the data-weight attribute on each. Now, loop over repLog of repLogs. Inside, set totalWeight += repLog.totalWeightLifted
It's nice to calculate the total weight from our source data, rather than reading it from somewhere on the DOM.
Okay! Try that out! The table still loads... and the total still prints!

 Tip

Actually, we made a mistake! When you delete a rep log, the total weight will no longer update! That's because we now need to remove the deleted repLog from the this.repLogs array.

No problem! The fix is kinda cool: it involves adding a reference to the $row element: the index on the this.repLogs array that the row corresponds to. This follows a pattern that's somewhat similar to what you'll see in ReactJS.