17.6.Array, Set and ES2016{Introducing Set}
But now, change the array to be a new Set(). To add items to a Set, you'll use add() instead of push() - but it's the same idea.
Try the script now.
Woh! Just two items! That's the key difference between Array and Set: Set should be used when you need a unique collection of items. It automatically makes sure that duplicates aren't added.
Oh, and there is also a WeakSet, which has the same super powers of WeakMap - all that garbage collection stuff. But, I haven't seen any decent use-case for it. Just use Set... or Array if values don't need to be unique.